# maestro-hive/sunday_com_backup_20251005/backend/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Sunday.com API
  description: |
    Comprehensive API documentation for Sunday.com work management platform.

    ## Authentication
    All API endpoints require authentication via JWT token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```

    ## Rate Limiting
    - 1000 requests per hour for authenticated users
    - 100 requests per hour for unauthenticated users

    ## Error Responses
    All error responses follow a consistent format:
    ```json
    {
      "error": {
        "type": "validation_error",
        "message": "Invalid request data",
        "details": {},
        "requestId": "uuid"
      }
    }
    ```

    ## Pagination
    List endpoints support cursor and offset-based pagination:
    ```
    GET /api/v1/boards?page=1&limit=20
    ```

  version: 1.0.0
  contact:
    name: Sunday.com Team
    email: api@sunday.com
    url: https://sunday.com/docs
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://sunday.com/terms
servers:
  - url: https://api.sunday.com/api/v1
    description: Production server
  - url: https://staging-api.sunday.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - bearerAuth: []

# ============================================================================
# AUTHENTICATION
# ============================================================================
paths:
...
  /organizations:
    get:
      tags: [Organizations]
      summary: List user's organizations
      description: Get all organizations where the user is a member
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Organizations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Organization'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags: [Organizations]
      summary: Create organization
      description: Create a new organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationRequest'
      responses:
        '201':
          description: Organization created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Organization'
        '422':
          $ref: '#/components/responses/ValidationError'
  /organizations/{organizationId}:
    get:
      tags: [Organizations]
      summary: Get organization
      description: Get organization details by ID
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      responses:
        '200':
          description: Organization retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Organization'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags: [Organizations]
      summary: Update organization
      description: Update organization details
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrganizationRequest'
      responses:
        '200':
          description: Organization updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Organization'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [Organizations]
      summary: Delete organization
      description: Soft delete an organization
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      responses:
        '204':
          description: Organization deleted successfully
        '404':
          $ref: '#/components/responses/NotFoundError'

# ============================================================================
# WORKSPACES
# ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Organization ID
...
  schemas:
...
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        domain:
          type: string
          nullable: true
        settings:
          type: object
          additionalProperties: true
        subscriptionPlan:
          type: string
          enum: [free, pro, enterprise]
        subscriptionStatus:
          type: string
          enum: [active, inactive, cancelled]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean

    # Request Schemas
    CreateOrganizationRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
          minLength: 3
          maxLength: 50
        domain:
          type: string
          format: hostname
        settings:
          type: object
          additionalProperties: true
    UpdateOrganizationRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        domain:
          type: string
          format: hostname
        settings:
          type: object
          additionalProperties: true
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [type, message]
          properties:
            type:
              type: string
              example: "validation_error"
            message:
              type: string
              example: "Invalid request data"
            details:
              type: object
              additionalProperties: true
            requestId:
              type: string
              format: uuid
  responses:
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              type: "not_found_error"
              message: "Resource not found"
              requestId: "req_123456"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              type: "validation_error"
              message: "Invalid request data"
              details:
                name: ["Name is required"]
                email: ["Invalid email format"]
              requestId: "req_123456"
...